+++ zcash_client_backend/src/data_api/wallet.rs
@@ -47,14 +47,16 @@ use super::InputSource;
 use crate::{
     data_api::{
         Account, MaxSpendMode, SentTransaction, SentTransactionOutput, WalletCommitmentTrees,
-        WalletRead, WalletWrite, error::Error, wallet::input_selection::propose_send_max,
+        WalletRead, WalletWrite,
+        error::{Error, LockError},
+        wallet::input_selection::propose_send_max,
     },
     decrypt_transaction,
     fees::{
         ChangeStrategy, DustOutputPolicy, StandardFeeRule, standard::SingleOutputChangeStrategy,
     },
     proposal::{Proposal, ProposalError, Step, StepOutputIndex},
-    wallet::{Note, OvkPolicy, Recipient},
+    wallet::{Note, OutputRef, OvkPolicy, Recipient},
 };
 use ::transparent::{
     address::TransparentAddress, builder::TransparentSigningSet, bundle::OutPoint,
@@ -603,6 +605,70 @@ impl ConfirmationsPolicy {
     }
 }
 
+/// Returns the set of [`OutputRef`]s for all inputs selected by the given proposal.
+fn proposal_input_refs<FeeRuleT, NoteRef>(
+    proposal: &Proposal<FeeRuleT, NoteRef>,
+) -> impl Iterator<Item = OutputRef> {
+    proposal.steps().iter().flat_map(|step| {
+        step.shielded_inputs()
+            .into_iter()
+            .flat_map(|shielded_inputs| {
+                shielded_inputs.notes().iter().map(|note| {
+                    let pool = match note.note() {
+                        Note::Sapling(_) => PoolType::SAPLING,
+                        #[cfg(feature = "orchard")]
+                        Note::Orchard(_) => PoolType::ORCHARD,
+                    };
+                    OutputRef::new(*note.txid(), pool, u32::from(note.output_index()))
+                })
+            })
+            .chain(step.transparent_inputs().iter().map(|utxo| {
+                let outpoint = utxo.outpoint();
+                OutputRef::new(
+                    TxId::from_bytes(*outpoint.hash()),
+                    PoolType::TRANSPARENT,
+                    outpoint.n(),
+                )
+            }))
+    })
+}
+
+/// Locks all inputs selected by the given proposal, preventing them from being
+/// selected by subsequent proposals. The lock expires at the given height.
+#[allow(clippy::type_complexity)]
+fn lock_proposal_inputs<DbT, FeeRuleT, NoteRef, TE, SE, FE, CE>(
+    wallet_db: &mut DbT,
+    proposal: &Proposal<FeeRuleT, NoteRef>,
+    lock_expiry_height: BlockHeight,
+) -> Result<(), Error<DbT::Error, TE, SE, FE, CE, NoteRef>>
+where
+    DbT: WalletWrite,
+{
+    match wallet_db.lock_outputs(proposal_input_refs(proposal), lock_expiry_height) {
+        Ok(_) => Ok(()),
+        Err(LockError::LockFailure(out_ref)) => {
+            Err(Error::Proposal(ProposalError::ChainDoubleSpend(out_ref)))
+        }
+        Err(LockError::Storage(e)) => Err(Error::DataSource(e)),
+    }
+}
+
+///
+/// This is useful when a proposal is rejected or abandoned after its inputs
+/// were locked via [`lock_proposal_inputs`].
+pub fn unlock_proposal_inputs<DbT, FeeRuleT, NoteRef>(
+    wallet_db: &mut DbT,
+    proposal: &Proposal<FeeRuleT, NoteRef>,
+) -> Result<(), DbT::Error>
+where
+    DbT: WalletWrite,
+{
+    for output_ref in proposal_input_refs(proposal) {
+        wallet_db.unlock_output(&output_ref)?;
+    }
+    Ok(())
+}
+
 /// Select transaction inputs, compute fees, and construct a proposal for a transaction or series
 /// of transactions that can then be authorized and made ready for submission to the network with
 /// [`create_proposed_transactions`].
@@ -616,13 +682,14 @@ pub fn propose_transfer<DbT, ParamsT, InputsT, ChangeT, CommitmentTreeErrT>(
     change_strategy: &ChangeT,
     request: zip321::TransactionRequest,
     confirmations_policy: ConfirmationsPolicy,
+    lock_for_blocks: Option<u32>,
     #[cfg(feature = "unstable")] proposed_version: Option<TxVersion>,
 ) -> Result<
     Proposal<ChangeT::FeeRule, <DbT as InputSource>::NoteRef>,
     ProposeTransferErrT<DbT, CommitmentTreeErrT, InputsT, ChangeT>,
 >
 where
-    DbT: WalletRead + InputSource<Error = <DbT as WalletRead>::Error>,
+    DbT: WalletWrite + InputSource<Error = <DbT as WalletRead>::Error>,
     <DbT as InputSource>::NoteRef: Copy + Eq + Ord,
     ParamsT: consensus::Parameters + Clone,
     InputsT: InputSelector<InputSource = DbT>,
@@ -650,6 +717,12 @@ where
         #[cfg(feature = "unstable")]
         proposed_version,
     )?;
+
+    if let Some(n) = lock_for_blocks {
+        let lock_expiry_height = target_height + n;
+        lock_proposal_inputs(wallet_db, &proposal, lock_expiry_height)?;
+    }
+
     Ok(proposal)
 }
 
@@ -691,6 +764,7 @@ pub fn propose_standard_transfer_to_address<DbT, ParamsT, CommitmentTreeErrT>(
     memo: Option<MemoBytes>,
     change_memo: Option<MemoBytes>,
     fallback_change_pool: ShieldedProtocol,
+    lock_for_blocks: Option<u32>,
     #[cfg(feature = "unstable")] proposed_version: Option<TxVersion>,
 ) -> Result<
     Proposal<StandardFeeRule, DbT::NoteRef>,
@@ -704,7 +778,10 @@ pub fn propose_standard_transfer_to_address<DbT, ParamsT, CommitmentTreeErrT>(
 where
     ParamsT: consensus::Parameters + Clone,
     DbT: InputSource,
-    DbT: WalletRead<Error = <DbT as InputSource>::Error, AccountId = <DbT as InputSource>::AccountId>,
+    DbT: WalletWrite<
+            Error = <DbT as InputSource>::Error,
+            AccountId = <DbT as InputSource>::AccountId,
+        >,
     DbT::NoteRef: Copy + Eq + Ord,
 {
     let request = zip321::TransactionRequest::new(vec![
@@ -738,6 +815,7 @@ where
         &change_strategy,
         request,
         confirmations_policy,
+        lock_for_blocks,
         #[cfg(feature = "unstable")]
         proposed_version,
     )
@@ -758,12 +836,13 @@ pub fn propose_send_max_transfer<DbT, ParamsT, FeeRuleT, CommitmentTreeErrT>(
     memo: Option<MemoBytes>,
     mode: MaxSpendMode,
     confirmations_policy: ConfirmationsPolicy,
+    lock_for_blocks: Option<u32>,
 ) -> Result<
     Proposal<FeeRuleT, <DbT as InputSource>::NoteRef>,
     ProposeSendMaxErrT<DbT, CommitmentTreeErrT, FeeRuleT>,
 >
 where
-    DbT: WalletRead + InputSource<Error = <DbT as WalletRead>::Error>,
+    DbT: WalletWrite + InputSource<Error = <DbT as WalletRead>::Error>,
     <DbT as InputSource>::NoteRef: Copy + Eq + Ord,
     ParamsT: consensus::Parameters + Clone,
     FeeRuleT: FeeRule + Clone,
@@ -791,6 +870,11 @@ where
         memo,
     )?;
 
+    if let Some(n) = lock_for_blocks {
+        let lock_expiry_height = target_height + n;
+        lock_proposal_inputs(wallet_db, &proposal, lock_expiry_height)?;
+    }
+
     Ok(proposal)
 }
 
@@ -808,13 +892,14 @@ pub fn propose_shielding<DbT, ParamsT, InputsT, ChangeT, CommitmentTreeErrT>(
     from_addrs: &[TransparentAddress],
     to_account: <DbT as InputSource>::AccountId,
     confirmations_policy: ConfirmationsPolicy,
+    lock_for_blocks: Option<u32>,
 ) -> Result<
     Proposal<ChangeT::FeeRule, Infallible>,
     ProposeShieldingErrT<DbT, CommitmentTreeErrT, InputsT, ChangeT>,
 >
 where
     ParamsT: consensus::Parameters,
-    DbT: WalletRead + InputSource<Error = <DbT as WalletRead>::Error>,
+    DbT: WalletWrite + InputSource<Error = <DbT as WalletRead>::Error>,
     InputsT: ShieldingSelector<InputSource = DbT>,
     ChangeT: ChangeStrategy<MetaSource = DbT>,
 {
@@ -823,7 +908,7 @@ where
         .map_err(|e| Error::from(InputSelectorError::DataSource(e)))?
         .ok_or_else(|| Error::from(InputSelectorError::SyncRequired))?;
 
-    input_selector
+    let proposal = input_selector
         .propose_shielding(
             params,
             wallet_db,
@@ -834,7 +919,14 @@ where
             (chain_tip_height + 1).into(),
             confirmations_policy,
         )
-        .map_err(Error::from)
+        .map_err(Error::from)?;
+
+    if let Some(n) = lock_for_blocks {
+        let lock_expiry_height = chain_tip_height + 1 + n;
+        lock_proposal_inputs(wallet_db, &proposal, lock_expiry_height)?;
+    }
+
+    Ok(proposal)
 }
 
 struct StepResult<AccountId> {
@@ -2612,6 +2704,7 @@ where
         from_addrs,
         to_account,
         confirmations_policy,
+        None,
     )?;
 
     create_proposed_transactions(