+++ zcash_client_backend/CHANGELOG.md
@@ -37,6 +37,7 @@ workspace.
 - `zcash_client_backend::data_api::error::LockError`
 - `zcash_client_backend::wallet::OutputRef`
 - `impl From<zcash_client_backend::wallet::NoteId> for zcash_client_backend::wallet::OutputRef`
+- `zcash_client_backend::data_api::wallet::unlock_proposal_inputs`
 
 ### Changed
 - `zcash_client_backend::data_api::wallet::create_proposed_transactions` now takes
@@ -96,6 +97,13 @@ workspace.
   `truncate_to_chain_state`, `lock_outputs` and `unlock_output`.
 - `zcash_client_backend::data_api::WalletTest` has added method
   `get_locked_outputs`.
+- The following `zcash_client_backend::data_api::wallet::` proposal creation
+  functions now take a `lock_for_blocks: Option<u32>` parameter and require
+  `WalletWrite` instead of `WalletRead`:
+  - `propose_transfer`
+  - `propose_standard_transfer_to_address`
+  - `propose_send_max_transfer`
+  - `propose_shielding`
 - The associated type `zcash_client_backend::fees::ChangeStrategy::MetaSource` is now
   bounded on the newly added `MetaSource` type instead of
   `zcash_client_backend::data_api::InputSource`.